A lot for navigation
I added javascript to changed class nojs to hasjs. I made the navigation pop out whenever the burger is clicked and retract when it is clicked again. Worked on the navigation so that the background would be all black and made a special condition for when the navigation is in the downloads page so that it goes down all the way to the bottom of the screen since that one is bigger than the index. I spent most of the day doing this. Making websites from scratch is painful yet rewarding. I'm going a little insane.