Registration and Download page mobile friendly
Finished making the Registration and Download page mobile friendly. This was accomplished by making the text bigger, spacing it out more and making the buttons bigger as well in the Downloads page. In the Registration Page I made all of the input boxes and labels bigger as well as make the checkbox and button bigger as well. I added some id's and deleted some html that I decided was unnecessary.